Distinguish between develop and latest configs
- Develop: may contain fixed versions or latest versions taken from
master. This config is supposed to be more stable than latest, and more
flexible in that it can be modified to test FCCSw against new versions
of the externals withouth deploying a a new release.
- Latest: is supposed to be highly unstable since it will build the
latest version of each package (might not be practical, if so, we will
remove it)